+++ Packages/com.unity.cloud.gltfast.tests/Tests/Runtime/Scripts/ImageFormatTests.cs
@@ -1,7 +1,9 @@
 // SPDX-FileCopyrightText: 2025 Unity Technologies and the glTFast authors
 // SPDX-License-Identifier: Apache-2.0
 
+using System.Text;
 using GLTFast.Export;
+using GLTFast.Schema;
 using NUnit.Framework;
 using UnityEngine;
 
@@ -12,14 +14,18 @@ namespace GLTFast.Tests
         [Test]
         public void FromMimeType()
         {
-            Assert.AreEqual(ImageFormat.Jpeg, ImageFormatExtensions.FromMimeType("image/jpeg"));
-            Assert.AreEqual(ImageFormat.Png, ImageFormatExtensions.FromMimeType("image/png"));
-            Assert.AreEqual(ImageFormat.Ktx, ImageFormatExtensions.FromMimeType("image/ktx"));
-            Assert.AreEqual(ImageFormat.Ktx, ImageFormatExtensions.FromMimeType("image/ktx2"));
-            Assert.AreEqual(ImageFormat.WebP, ImageFormatExtensions.FromMimeType("image/webp"));
-            Assert.AreEqual(ImageFormat.Unknown, ImageFormatExtensions.FromMimeType("image/fantasy-format"));
-            Assert.AreEqual(ImageFormat.Unknown, ImageFormatExtensions.FromMimeType("application/jpeg"));
-            Assert.AreEqual(ImageFormat.Unknown, ImageFormatExtensions.FromMimeType(null));
+            Assert.AreEqual(ImageFormat.Jpeg, ImageFormatExtensions.FromMimeType(ImageMimeType.Jpeg));
+            Assert.AreEqual(ImageFormat.Png, ImageFormatExtensions.FromMimeType(ImageMimeType.Png));
+            Assert.AreEqual(ImageFormat.Ktx, ImageFormatExtensions.FromMimeType(ImageMimeType.Ktx2));
+            Assert.AreEqual(ImageFormat.WebP, ImageFormatExtensions.FromMimeType(ImageMimeType.WebP));
+            Assert.AreEqual(ImageFormat.Unknown, ImageFormatExtensions.FromMimeType(ImageMimeType.Undefined));
+            Assert.AreEqual(ImageFormat.Unknown, ImageFormatExtensions.FromMimeType(default(EnumOrRawValue<ImageMimeType>)));
+            Assert.AreEqual(ImageFormat.Unknown, ImageFormatExtensions.FromMimeType(
+                new EnumOrRawValue<ImageMimeType>(Encoding.UTF8.GetBytes("image/ktx"))));
+            Assert.AreEqual(ImageFormat.Unknown, ImageFormatExtensions.FromMimeType(
+                new EnumOrRawValue<ImageMimeType>(Encoding.UTF8.GetBytes("image/fantasy-format"))));
+            Assert.AreEqual(ImageFormat.Unknown, ImageFormatExtensions.FromMimeType(
+                new EnumOrRawValue<ImageMimeType>(Encoding.UTF8.GetBytes("application/jpeg"))));
         }
     }
 }