+++ Packages/com.unity.cloud.gltfast/Runtime/Scripts/Export/Constants.cs
@@ -8,8 +8,5 @@ namespace GLTFast.Export
     static class Constants
     {
         public const string version = "7.0.0-exp.1";
-
-        internal const string mimeTypePNG = "image/png";
-        internal const string mimeTypeJPG = "image/jpeg";
     }
 }