+++ Packages/com.unity.cloud.gltfast/Runtime/Scripts/Export/MaskMapImageExport.cs
@@ -25,7 +25,7 @@ namespace GLTFast.Export
             : base(maskMap, imageFormat) { }
 
         /// <inheritdoc />
-        protected override ImageFormat ImageFormat => m_ImageFormat != ImageFormat.Unknown ? m_ImageFormat : ImageFormat.Jpg;
+        protected override ImageFormat ImageFormat => m_ImageFormat != ImageFormat.Unknown ? m_ImageFormat : ImageFormat.Jpeg;
 
         static Material GetMaskMapBlitMaterial()
         {