+++ Packages/com.unity.cloud.gltfast/Runtime/Scripts/Export/OrmImageExport.cs
@@ -68,7 +68,7 @@ namespace GLTFast.Export
         }
 
         /// <inheritdoc />
-        protected override ImageFormat ImageFormat => m_ImageFormat != ImageFormat.Unknown ? m_ImageFormat : ImageFormat.Jpg;
+        protected override ImageFormat ImageFormat => m_ImageFormat != ImageFormat.Unknown ? m_ImageFormat : ImageFormat.Jpeg;
 
         /// <inheritdoc />
         public override FilterMode FilterMode